kyc minor refactor
